Refactor
Big refactoring of the current navigation. I also added "change email" function and it works (with Firebase Authentication).

Overall a lot of changes applied. Applied the ClassDiagram structure to the project for Login / Register / Main and Profile (so far I guess).

I also merged development into register, to be up-to-date with the latest changes made into development.